requirements.txt
requests==2.26.0
paho-mqtt==1.5.1
schedule==1.1.0